Refactor src dirs
organize sources to logical components
