+++ src/converter/med/common_med.cc
@@ -3,10 +3,10 @@
 
 #include <vector>
 #include "common_med.h"
-#include "converter.h"
+#include "converter/converter.h"
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wunused-result"
-#include "../3rdparty/SimpleBinStream.h"
+#include "3rdparty/SimpleBinStream.h"
 #pragma GCC diagnostic pop
 
 using namespace converter::med;
