+++ src/converter/med/converter_med4.cc
@@ -9,8 +9,8 @@
 #include <string>
 #include <utility>
 #include <vector>
-#include "big_endian.h"
-#include "converter.h"
+#include "converter/big_endian.h"
+#include "converter/converter.h"
 #include "common_med.h"
 
 using namespace std;
