+++ src/plugin/audacious/prefs.h
