+++ src/plugin/audacious/uade_config.h
