+++ src/uade/hacks.cc
