+++ src/uade/hacks.h
