+++ src/uade/uade_common.h
